requirements.txt
requests==2.25.1
twitch-python==0.0.19
